refactor: Migrate test.ts
- Removed namespaces
- Added imports
- @TODO: DOcumentation
- @TODO: Check if code needs to be moved
